requirements.txt
click==8.0.1
cloudinary==1.26.0
dnspython==2.1.0
Flask==2.0.1
<<<<<<< HEAD
=======
Flask-Cors==3.0.10
>>>>>>> 8054a8d96c3c7b315e036728df8177d5cb6a674f
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
itsdangerous==2.0.1
pymongo==3.12.0
<<<<<<< HEAD
=======
python-dotenv==0.19.0
>>>>>>> 8054a8d96c3c7b315e036728df8177d5cb6a674f
Werkzeug==2.0.1
